Extract utilities into
