+++ src/utils/calculateWidth.js
@@ -1,0 +1,5 @@
+const WIDTH = 82;
+
+const calculateWidth = (duration) => (WIDTH * duration) - 2;
+
+export default calculateWidth;
